Revert "cmd/atlas/internal: allow  extension of migratr command (#1904)"
This reverts commit 08c3ff985edbaf3b7aaa4b6fab04f2525ef32964.
