+++ cmd/atlas/internal/cmdapi/migrate.go
@@ -41,6 +41,22 @@ import (
 	"github.com/spf13/cobra"
 )
 
+func init() {
+	migrateCmd := migrateCmd()
+	migrateCmd.AddCommand(
+		migrateApplyCmd(),
+		migrateDiffCmd(),
+		migrateHashCmd(),
+		migrateImportCmd(),
+		migrateLintCmd(),
+		migrateNewCmd(),
+		migrateSetCmd(),
+		migrateStatusCmd(),
+		migrateValidateCmd(),
+	)
+	Root.AddCommand(migrateCmd)
+}
+
 // migrateCmd represents the subcommand 'atlas migrate'.
 func migrateCmd() *cobra.Command {
 	cmd := &cobra.Command{
@@ -507,11 +523,6 @@ type migrateDiffFlags struct {
 
 // migrateDiffCmd represents the 'atlas migrate diff' subcommand.
 func migrateDiffCmd() *cobra.Command {
-	cmd, _ := migrateDiffWithFlagsCmd()
-	return cmd
-}
-
-func migrateDiffWithFlagsCmd() (*cobra.Command, *migrateDiffFlags) {
 	var (
 		flags migrateDiffFlags
 		cmd   = &cobra.Command{
@@ -556,7 +567,7 @@ directory state to the desired schema. The desired state can be another connecte
 	cmd.Flags().BoolVarP(&flags.edit, flagEdit, "", false, "edit the generated migration file(s)")
 	cobra.CheckErr(cmd.MarkFlagRequired(flagTo))
 	cobra.CheckErr(cmd.MarkFlagRequired(flagDevURL))
-	return cmd, &flags
+	return cmd
 }
 
 func migrateDiffRun(cmd *cobra.Command, args []string, flags migrateDiffFlags, env *Env) error {
