--- cmd/atlas/internal/cmdapi/migrate_oss.go
@@ -1,23 +1,0 @@
-// Copyright 2021-present The Atlas Authors. All rights reserved.
-// This source code is licensed under the Apache 2.0 license found
-// in the LICENSE file in the root directory of this source tree.
-
-//go:build !ent
-
-package cmdapi
-
-func init() {
-	migrateCmd := migrateCmd()
-	migrateCmd.AddCommand(
-		migrateApplyCmd(),
-		migrateDiffCmd(),
-		migrateHashCmd(),
-		migrateImportCmd(),
-		migrateLintCmd(),
-		migrateNewCmd(),
-		migrateSetCmd(),
-		migrateStatusCmd(),
-		migrateValidateCmd(),
-	)
-	Root.AddCommand(migrateCmd)
-}
